New organizational scheme for log files
Log files for Tensorboard are now stored under one master dir in the root of ML_Indiana instead of under 'runs'. this allows them to be compared directly to one another in a TensorBoard session. Tensorboard logs are named identically to the run they correspond with. Removed chi sqaured analysis from process as it seem misused in this case.